Extracted dependent views helpers to share between migrations
